+++ vkconfig_core/test/test_layer.cpp
@@ -36,7 +36,7 @@ static SettingMetaString* InstantiateString(Layer& layer, const std::string& key
     return static_cast<SettingMetaString*>(layer.Instantiate(layer.settings, key, SETTING_STRING));
 }
 
-static std::map<Path, std::string> Dummy() { return std::map<Path, std::string>(); }
+static std::map<Path, LayerStatus> Dummy() { return std::map<Path, LayerStatus>(); }
 
 TEST(test_layer, collect_settings) {
     Layer layer;